OSUCTFBootcamp/Crypto: Number Theory/mpaste/chall/app/requirements.txt
click==7.1.2
Faker==5.3.0
Flask==1.1.2
Flask-Limiter==1.4
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
pycryptodome==3.9.9
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
text-unidecode==1.3
uWSGI==2.0.19.1
Werkzeug==1.0.1
